RHCLOUD-5248 break down sat-receptor response handlers into individual files
